+++ Utilities/PrimerMathUtils.cs
@@ -4,7 +4,7 @@ namespace PrimerTools;
 
 public static class PrimerMathUtils
 {
-    public static float DecayingExponentialWithMin(float initialDuration, float decayFactor, float minDuration, float step)
+    public static double DecayingExponentialWithMin(double initialDuration, double decayFactor, double minDuration, double step)
     {
         // TODO: Could fit this to an actual exponential instead of clamping, but who cares
         return Mathf.Max(initialDuration * Mathf.Pow(decayFactor, step), minDuration);
